requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
python-dateutil==2.8.2
google-generativeai>=0.3.0

